requirements.txt
# M.SAVAŞ Video Analiz Sistemi - Gerekli Kütüphaneler
# Python 3.8+ gereklidir

# Temel GUI framework
PyQt5>=5.15.0

# Bilgisayarlı görü ve video işleme
opencv-python>=4.5.0
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0

# Rapor oluşturma kütüphaneleri
xlsxwriter>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.0

# Grafik ve görselleştirme
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=8.3.0

# Sayısal işlemler
numpy>=1.21.0
scipy>=1.7.0

# Yardımcı kütüphaneler
psutil>=5.8.0
tqdm>=4.62.0
